+++ spec/active_record/postgres_pub_sub/listener_spec.rb
@@ -9,7 +9,8 @@ RSpec.describe ActiveRecord::PostgresPubSub::Listener, cleaner_strategy: :trunca
       OpenStruct.new(started: 0,
                      count: 0,
                      timeout_count: 0,
-                     payloads: [])
+                     payloads: [],
+                     received_channels: [])
     end
     let!(:listener_thread) do
       Thread.new do
@@ -34,6 +35,33 @@ RSpec.describe ActiveRecord::PostgresPubSub::Listener, cleaner_strategy: :trunca
       wait_for("notification received") { state.count > 0 }
       expect(state.payloads).to match_ordered_array([nil])
       expect(state.count).to eq(1)
+    end
+
+    context "when using 1-arg version of #on_notify" do
+      let!(:listener_thread) do
+        Thread.new do
+          listener_loop(listener_options) do |listener|
+            listener.on_notify do |payload|
+              state.count += 1
+              state.payloads << payload
+            end
+          end
+        ensure
+          ActiveRecord::Base.clear_active_connections!
+        end
+      end
+
+      it "invokes the single arg. notify block when it receives a notification" do
+        wait_for_started
+
+        ActiveRecord::Base.transaction do
+          3.times { |i| notify(i) }
+        end
+
+        wait_for("notification received") { state.count > 0 }
+        expect(state.payloads).to match_ordered_array([nil])
+        expect(state.count).to eq(1)
+      end
     end
 
     context "when using notify_only=false" do
@@ -65,9 +93,26 @@ RSpec.describe ActiveRecord::PostgresPubSub::Listener, cleaner_strategy: :trunca
       end
     end
 
-    def notify(payload)
+    context "when listen to multiple channels" do
+      let(:channels) { %w(pub_sub_test1 pub_sub_test2) }
+      let(:listener_options) { Hash[listen_to: channels, notify_only: false] }
+
+      it "invokes the notify multiple channels block with notification notify to diffrent channels" do
+        wait_for_started
+
+        ActiveRecord::Base.transaction do
+          channels.each { |c| notify(c, notify_to: c) }
+        end
+
+        wait_for("notification received") { state.received_channels.count > 0 }
+        expect(state.payloads).to match_ordered_array(channels)
+        expect(state.received_channels).to match_ordered_array(channels)
+      end
+    end
+
+    def notify(payload, notify_to: channel)
       # rubocop:disable Ezcater/RailsTopLevelSqlExecute
-      ActiveRecord::Base.connection.execute("NOTIFY #{channel}, '#{payload}'")
+      ActiveRecord::Base.connection.execute("NOTIFY #{notify_to}, '#{payload}'")
       # rubocop:enable Ezcater/RailsTopLevelSqlExecute
     end
 
@@ -85,8 +130,8 @@ RSpec.describe ActiveRecord::PostgresPubSub::Listener, cleaner_strategy: :trunca
       end
     end
 
-    def listener_loop(listen_timeout: nil, exclusive_lock: true, notify_only: true)
-      described_class.listen(channel,
+    def listener_loop(listen_to: [channel], listen_timeout: nil, exclusive_lock: true, notify_only: true)
+      described_class.listen(*listen_to,
                              listen_timeout: listen_timeout,
                              exclusive_lock: exclusive_lock,
                              notify_only: notify_only) do |listener|
@@ -94,14 +139,17 @@ RSpec.describe ActiveRecord::PostgresPubSub::Listener, cleaner_strategy: :trunca
           state.started += 1
         end
 
-        listener.on_notify do |payload|
+        listener.on_notify do |payload, channel|
           state.count += 1
           state.payloads << payload
+          state.received_channels << channel
         end
 
         listener.on_timeout do
           state.timeout_count += 1
         end
+
+        yield(listener) if block_given?
       end
     end
   end